08_10_24/requirements.txt
numpy
pandas
faker
gspread
google-auth